解决bug，已完任务数量显示错误，当恢复已完成的任务后，任务数量还显示原来的。

+++ app/src/main/java/com/ninestone/morefficient/view/v/DoneTaskView.java
@@ -18,5 +18,5 @@ public interface DoneTaskView {
      * 填充任务数量
      * @param count
      */
-    void fillCount(int count);
+    void fillCount(long count);
 }
